- Delete and decouple gillespy2.py - All imports are handled in __init__.py - import gillespy2 works the same as previously
