+++ gillespy2/core/Utilities.py
@@ -1,0 +1,400 @@
+from gillespy2 import *
+from collections import OrderedDict
+
+try:
+    import lxml.etree as eTree
+
+    no_pretty_print = False
+
+except:
+    import xml.etree.ElementTree as eTree
+    import xml.dom.minidom
+    import re
+    no_pretty_print = True
+
+def import_SBML(filename, name=None, gillespy_model=None):
+    """
+    SBML to GillesPy model converter. NOTE: non-mass-action rates
+    in terms of concentrations may not be converted for population
+    simulation. Use caution when importing SBML.
+
+    Attributes
+    ----------
+    filename : str
+        Path to the SBML file for conversion.
+    name : str
+        Name of the resulting model.
+    gillespy_model : gillespy.Model
+        If desired, the SBML model may be added to an existing GillesPy model.
+    """
+
+    try:
+        from gillespy2.sbml.SBMLimport import convert
+    except ImportError:
+        raise ImportError('SBML conversion not imported successfully')
+
+    return convert(filename, model_name=name, gillespy_model=gillespy_model)
+
+
+class StochMLDocument():
+    """ Serializiation and deserialization of a Model to/from
+        the native StochKit2 XML format. """
+
+    def __init__(self):
+        # The root element
+        self.document = eTree.Element("Model")
+        self.annotation = None
+
+    @classmethod
+    def from_model(cls, model):
+        """
+        Creates an StochKit XML document from an exisiting Mdoel object.
+        This method assumes that all the parameters in the model are already
+        resolved to scalar floats (see Model.resolveParamters).
+
+        Note, this method is intended to be used interanally by the models
+        'serialization' function, which performs additional operations and
+        tests on the model prior to writing out the XML file. You should NOT \
+        do:
+
+        document = StochMLDocument.fromModel(model)
+        print document.toString()
+
+        You SHOULD do
+
+        print model.serialize()
+
+        """
+
+        # Description
+        md = cls()
+
+        d = eTree.Element('Description')
+
+        #
+        if model.units.lower() == "concentration":
+            d.set('units', model.units.lower())
+
+        d.text = model.annotation
+        md.document.append(d)
+
+        # Number of Reactions
+        nr = eTree.Element('NumberOfReactions')
+        nr.text = str(len(model.listOfReactions))
+        md.document.append(nr)
+
+        # Number of Species
+        ns = eTree.Element('NumberOfSpecies')
+        ns.text = str(len(model.listOfSpecies))
+        md.document.append(ns)
+
+        # Species
+        spec = eTree.Element('SpeciesList')
+        for sname in model.listOfSpecies:
+            spec.append(md.__species_to_element(model.listOfSpecies[sname]))
+        md.document.append(spec)
+
+        # Parameters
+        params = eTree.Element('ParametersList')
+        for pname in model.listOfParameters:
+            params.append(md.__parameter_to_element(
+                model.listOfParameters[pname]))
+
+        params.append(md.__parameter_to_element(Parameter(name='vol', expression=model.volume)))
+
+        md.document.append(params)
+
+        # Reactions
+        reacs = eTree.Element('ReactionsList')
+        for rname in model.listOfReactions:
+            reacs.append(md.__reaction_to_element(model.listOfReactions[rname], model.volume))
+        md.document.append(reacs)
+
+        return md
+
+    @classmethod
+    def from_file(cls, filepath):
+        """ Intializes the document from an exisiting native StochKit XML
+        file read from disk. """
+        tree = eTree.parse(filepath)
+        root = tree.getroot()
+        md = cls()
+        md.document = root
+        return md
+
+    @classmethod
+    def from_string(cls, string):
+        """ Intializes the document from an exisiting native StochKit XML
+        file read from disk. """
+        root = eTree.fromString(string)
+
+        md = cls()
+        md.document = root
+        return md
+
+    def to_model(self, name):
+        """ Instantiates a Model object from a StochMLDocument. """
+
+        # Empty model
+        model = Model(name=name)
+        root = self.document
+
+        # Try to set name from document
+        if model.name == "":
+            name = root.find('Name')
+            if name.text is None:
+                raise NameError("The Name cannot be none")
+            else:
+                model.name = name.text
+
+        # Set annotiation
+        ann = root.find('Description')
+        if ann is not None:
+            units = ann.get('units')
+
+            if units:
+                units = units.strip().lower()
+
+            if units == "concentration":
+                model.units = "concentration"
+            elif units == "population":
+                model.units = "population"
+            else:  # Default
+                model.units = "population"
+
+            if ann.text is None:
+                model.annotation = ""
+            else:
+                model.annotation = ann.text
+
+        # Set units
+        units = root.find('Units')
+        if units is not None:
+            if units.text.strip().lower() == "concentration":
+                model.units = "concentration"
+            elif units.text.strip().lower() == "population":
+                model.units = "population"
+            else:  # Default
+                model.units = "population"
+
+        # Create parameters
+        for px in root.iter('Parameter'):
+            name = px.find('Id').text
+            expr = px.find('Expression').text
+            if name.lower() == 'vol' or name.lower() == 'volume':
+                model.volume = float(expr)
+            else:
+                p = Parameter(name, expression=expr)
+                # Try to evaluate the expression in the empty namespace
+                # (if the expr is a scalar value)
+                p.evaluate()
+                model.add_parameter(p)
+
+        # Create species
+        for spec in root.iter('Species'):
+            name = spec.find('Id').text
+            val = spec.find('InitialPopulation').text
+            if '.' in val:
+                val = float(val)
+            else:
+                val = int(val)
+            s = Species(name, initial_value=val)
+            model.add_species([s])
+
+        # The namespace_propensity for evaluating the propensity function
+        # for reactions must contain all the species and parameters.
+        namespace_propensity = OrderedDict()
+        all_species = model.get_all_species()
+        all_parameters = model.get_all_parameters()
+
+        for param in all_species:
+            namespace_propensity[param] = all_species[param].initial_value
+
+        for param in all_parameters:
+            namespace_propensity[param] = all_parameters[param].value
+
+        # Create reactions
+        for reac in root.iter('Reaction'):
+            try:
+                name = reac.find('Id').text
+            except:
+                raise InvalidStochMLError("Reaction has no name.")
+
+            reaction = Reaction(name=name, reactants={}, products={})
+
+            # Type may be 'mass-action','customized'
+            try:
+                type = reac.find('Type').text
+            except:
+                raise InvalidStochMLError("No reaction type specified.")
+
+            reactants = reac.find('Reactants')
+            try:
+                for ss in reactants.iter('SpeciesReference'):
+                    specname = ss.get('id')
+                    # The stochiometry should be an integer value, but some
+                    # exising StoxhKit models have them as floats. This is
+                    # why we need the slightly odd conversion below.
+                    stoch = int(float(ss.get('stoichiometry')))
+                    # Select a reference to species with name specname
+                    sref = model.listOfSpecies[specname]
+                    try:
+                        # The sref list should only contain one element if
+                        # the XML file is valid.
+                        reaction.reactants[sref] = stoch
+                    except Exception as e:
+                        StochMLImportError(e)
+            except:
+                # Yes, this is correct. 'reactants' can be None
+                pass
+
+            products = reac.find('Products')
+            try:
+                for ss in products.iter('SpeciesReference'):
+                    specname = ss.get('id')
+                    stoch = int(float(ss.get('stoichiometry')))
+                    sref = model.listOfSpecies[specname]
+                    try:
+                        # The sref list should only contain one element if
+                        # the XML file is valid.
+                        reaction.products[sref] = stoch
+                    except Exception as e:
+                        raise StochMLImportError(e)
+            except:
+                # Yes, this is correct. 'products' can be None
+                pass
+
+            if type == 'mass-action':
+                reaction.massaction = True
+                reaction.type = 'mass-action'
+                # If it is mass-action, a parameter reference is needed.
+                # This has to be a reference to a species instance. We
+                # explicitly disallow a scalar value to be passed as the
+                # parameter.
+                try:
+                    ratename = reac.find('Rate').text
+                    try:
+                        reaction.marate = model.listOfParameters[ratename]
+                    except KeyError as k:
+                        # No paramter name is given. This is a valid use case
+                        # in StochKit. We generate a name for the paramter,
+                        # and create a new parameter instance. The parameter's
+                        # value should now be found in 'ratename'.
+                        generated_rate_name = "Reaction_" + name + \
+                                              "_rate_constant"
+                        p = Parameter(name=generated_rate_name,
+                                      expression=ratename)
+                        # Try to evaluate the parameter to set its value
+                        p.evaluate()
+                        model.add_parameter(p)
+                        reaction.marate = model.listOfParameters[
+                            generated_rate_name]
+
+                    reaction.__create_mass_action()
+                except Exception as e:
+                    raise
+            elif type == 'customized':
+                try:
+                    propfunc = reac.find('PropensityFunction').text
+                except Exception as e:
+                    raise InvalidStochMLError(
+                        "Found a customized propensity function, but no expression was given. {}".format(e))
+                reaction.propensity_function = propfunc
+            else:
+                raise InvalidStochMLError(
+                    "Unsupported or no reaction type given for reaction" + name)
+
+            model.add_reaction(reaction)
+
+        return model
+
+    def to_string(self):
+        """ Returns  the document as a string. """
+        try:
+            doc = eTree.tostring(self.document, pretty_print=True)
+            return doc.decode("utf-8")
+        except:
+            # Hack to print pretty xml without pretty-print
+            # (requires the lxml module).
+            doc = eTree.tostring(self.document)
+            xmldoc = xml.dom.minidom.parseString(doc)
+            uglyXml = xmldoc.toprettyxml(indent='  ')
+            text_re = re.compile(">\n\s+([^<>\s].*?)\n\s+</", re.DOTALL)
+            prettyXml = text_re.sub(">\g<1></", uglyXml)
+            return prettyXml
+
+    def __species_to_element(self, S):
+        e = eTree.Element('Species')
+        idElement = eTree.Element('Id')
+        idElement.text = S.name
+        e.append(idElement)
+
+        if hasattr(S, 'description'):
+            descriptionElement = eTree.Element('Description')
+            descriptionElement.text = S.description
+            e.append(descriptionElement)
+
+        initialPopulationElement = eTree.Element('InitialPopulation')
+        initialPopulationElement.text = str(S.initial_value)
+        e.append(initialPopulationElement)
+
+        return e
+
+    def __parameter_to_element(self, P):
+        e = eTree.Element('Parameter')
+        idElement = eTree.Element('Id')
+        idElement.text = P.name
+        e.append(idElement)
+        expressionElement = eTree.Element('Expression')
+        expressionElement.text = str(P.value)
+        e.append(expressionElement)
+        return e
+
+    def __reaction_to_element(self, R, model_volume):
+        e = eTree.Element('Reaction')
+
+        idElement = eTree.Element('Id')
+        idElement.text = R.name
+        e.append(idElement)
+
+        descriptionElement = eTree.Element('Description')
+        descriptionElement.text = self.annotation
+        e.append(descriptionElement)
+
+        # StochKit2 wants a rate for mass-action propensites
+        if R.massaction and model_volume == 1.0:
+            rateElement = eTree.Element('Rate')
+            # A mass-action reactions should only have one parameter
+            rateElement.text = R.marate.name
+            typeElement = eTree.Element('Type')
+            typeElement.text = 'mass-action'
+            e.append(typeElement)
+            e.append(rateElement)
+
+        else:
+            typeElement = eTree.Element('Type')
+            typeElement.text = 'customized'
+            e.append(typeElement)
+            functionElement = eTree.Element('PropensityFunction')
+            functionElement.text = R.propensity_function
+            e.append(functionElement)
+
+        reactants = eTree.Element('Reactants')
+
+        for reactant, stoichiometry in R.reactants.items():
+            srElement = eTree.Element('SpeciesReference')
+            srElement.set('id', str(reactant))
+            srElement.set('stoichiometry', str(stoichiometry))
+            reactants.append(srElement)
+
+        e.append(reactants)
+
+        products = eTree.Element('Products')
+        for product, stoichiometry in R.products.items():
+            srElement = eTree.Element('SpeciesReference')
+            srElement.set('id', str(product))
+            srElement.set('stoichiometry', str(stoichiometry))
+            products.append(srElement)
+        e.append(products)
+
+        return e
